feat(rename-switch): 开关重命名代码优化
