+++ src/components/Modal/Modal.less
@@ -101,14 +101,14 @@
         font-size: 32px;
         border-right: 1px solid #E7E8EB;
         &.no-border {
-          border: none; 
+          border: none;
         }
 
         &:last-of-type {
           border-right: none;
         }
       }
-    } 
+    }
 
     .btn-group.confirm-btn-group {
       padding-top: 16px;
